Add eglib library from the mono project
This adds a deps/ulib api which is a small MIT licensed library that is
api compatible with enough of the glib api to run Cogl. This replaces
deps/glib and deps/gmodule for use when Cogl is built with
--enable-standalone.

This code is based on eglib from the mono project, with a few updates
and additions to handle apis that Cogl uses that mono doesn't.

As a reference point for monitoring eglib changes in the mono project,
this code is based on commit d50c836721f22df74a45185d51d0b9d37c370e33